# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexora API",
    "version": "1.0.0",
    "description": "REST API for Nexora, a social networking platform built with Node.js, Express.js, ES Modules, JWT authentication, Multer file uploads, and in-memory data storage."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/posts/all": {
      "get": {
        "tags": ["Posts"],
        "summary": "Retrieve all published posts",
        "description": "Retrieves published posts with optional caption filtering, sorting, and pagination. Posts are sorted by date by default.",
        "operationId": "getAllPosts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "caption",
            "in": "query",
            "required": false,
            "description": "Filter posts whose captions contain the supplied text",
            "schema": {
              "type": "string"
            },
            "example": "first post"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "Sort posts by creation date or engagement score",
            "schema": {
              "type": "string",
              "enum": ["date", "engagement"]
            },
            "example": "engagement"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of posts per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Retrieve posts belonging to the authenticated user",
        "operationId": "getAuthenticatedUserPosts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPostListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create a new post",
        "operationId": "createPost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["caption"],
                "properties": {
                  "caption": {
                    "type": "string",
                    "maxLength": 2200,
                    "description": "Post caption"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional JPEG, PNG, or WebP image with a maximum size of 5 MB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/PostId"
        }
      ],
      "get": {
        "tags": ["Posts"],
        "summary": "Retrieve a specific post",
        "operationId": "getPostById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/PostNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": ["Posts"],
        "summary": "Update a post",
        "description": "Updates a post belonging to the authenticated user. At least one of caption or image must be supplied.",
        "operationId": "updatePost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "caption": {
                    "type": "string",
                    "maxLength": 2200
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG, or WebP image with a maximum size of 5 MB"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/PostNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete a post",
        "operationId": "deletePost",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "status": "success",
                  "message": "Post deleted successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/PostNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/posts/{id}/status": {
      "patch": {
        "tags": ["Posts"],
        "summary": "Update post status",
        "description": "Changes a post status to published, draft, or archived.",
        "operationId": "updatePostStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PostId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostSuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/PostNotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token returned by POST /api/signin"
      }
    },
    "parameters": {
      "PostId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Post UUID",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "schemas": {
...
      "Post": {
        "type": "object",
        "required": ["id", "userId", "caption", "status", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "caption": {
            "type": "string",
            "maxLength": 2200
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["published", "draft", "archived"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PostStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["published", "draft", "archived"]
          }
        }
      },
...
      "PostPagination": {
        "type": "object",
        "required": ["currentPage", "limit", "totalPosts", "totalPages"],
        "properties": {
          "currentPage": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "totalPosts": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
...
      "PostListResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/PostPagination"
              }
            }
          }
        }
      },
      "UserPostListResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      },
      "PostResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "post": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        }
      },
      "PostSuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "post": {
                "$ref": "#/components/schemas/Post"
              }
            }
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ValidationErrorItem": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["status", "message", "errors"],
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Authentication failed or authentication token was not supplied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Authentication required"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Authenticated user is not authorized to perform the operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "You are not authorized to perform this operation"
            }
          }
        }
      },
      "PostNotFoundError": {
        "description": "Post not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "fail",
              "message": "Post not found"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": "error",
              "message": "Internal server error"
            }
          }
        }
      }
    }
  }
}